git-svn-id: http://jrt.choury.com/svn/myos@44 2cdf69c7-79f9-4451-ab89-d57ae126d6d6

+++ kernel/syscall.c
@@ -15,17 +15,17 @@ int syscall(u32 eax,u32 ebx,u32 ecx,u32 edx,u32 esi,u32 edi){
     switch(eax){
     case 1:
         if(!reenter){
-            ecx+=PROTABLE[curpid].base;
+            ecx+=(u32)PROTABLE[curpid].base;
         }
         return sys_write((int)ebx,(const void *)ecx,(size_t)edx);
     case 2:
         if(!reenter){
-            ecx+=PROTABLE[curpid].base;
+            ecx+=(u32)PROTABLE[curpid].base;
         }
         return sys_read((int)ebx,(void *)ecx,(size_t)edx);
     case 3:
         if(!reenter){
-            ebx+=PROTABLE[curpid].base;
+            ebx+=(u32)PROTABLE[curpid].base;
         }
         return sys_open((const char*)ebx,(int)ecx,edx);
     case 4:
@@ -38,15 +38,15 @@ int syscall(u32 eax,u32 ebx,u32 ecx,u32 edx,u32 esi,u32 edi){
         return sys_lseek((int)ebx,(off_t)ecx,(int)edx);
     case 8:
         if(!reenter){
-            ebx+=PROTABLE[curpid].base;
-            ecx+=PROTABLE[curpid].base;
+            ebx+=(u32)PROTABLE[curpid].base;
+            ecx+=(u32)PROTABLE[curpid].base;
         }
         return sys_gettimeofday((struct timeval *)ebx,(struct timezone *)ecx);
     case 9:
         if(!reenter){
-            ebx+=PROTABLE[curpid].base;
-            ecx+=PROTABLE[curpid].base;
-            edx+=PROTABLE[curpid].base;
+            ebx+=(u32)PROTABLE[curpid].base;
+            ecx+=(u32)PROTABLE[curpid].base;
+            edx+=(u32)PROTABLE[curpid].base;
         }
         return sys_execve((char *)ebx,(char **)ecx,(char **)edx);
     case 10:
